+++ packages/client/src/pages/ProfileChangePwd/ProfileChangePwd.tsx
@@ -5,13 +5,13 @@ import { T_ProfileSchema, I_UserInfoData, I_UserPwdData, validationProfileSchema
 import { useNavigate } from 'react-router-dom';
 import { useNotification } from '../../global/hooks';
 import { UseFormReturn } from 'react-hook-form';
-import { PATHS } from '../../routes';
 import { userApi } from '../../api/User';
 import classes from './ProfileChangePwd.module.scss';
 import { authApi } from '../../api';
 import profileBG from '../../assets/images/game-main-menu-bg.jpg';
 import { Button } from '@mui/material';
 import { NavLink } from 'react-router-dom';
+import { PATHS } from '../../global/paths';
 
 const FIELDS: T_FormFieldNames = ['oldPassword', 'newPassword', 'confirmPassword'];
